pyproject.toml
[project]
name = "webroot"
version = "1.0.0"
description = "Discord Bot"
authors = [
    {name = "CYRAX",email = "cherniq66@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "disnake (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
